create server folder, start to refactor
