+++ desktop/src-tauri/Cargo.toml
@@ -38,16 +38,10 @@ serde = { version = "^1.0.198", features = ["derive"] }
 url = "2.5.0"
 open = "5.1.3"
 showfile = "0.1.1"
-clap = { version = "4.5.4", features = ["derive", "string"] }
 futures = "0.3.30"
 cpal = "0.15.3"
 hound = "3.5.1"
 rand = "0.8.5"
-
-# Server
-axum = { version = "0.7.5", optional = true }
-utoipa = { version = "4.2.3", features = ["axum_extras"], optional = true }
-utoipa-swagger-ui = { version = "7.1.0", features = ["axum"], optional = true }
 
 # Log
 tracing = { version = "0.1.40", features = ["log"] }
@@ -59,43 +53,18 @@ chrono = "0.4.38"
 crash-handler = "0.6.2"
 urlencoding = "2.1.3"
 
-
 # Linux
 [target.'cfg(target_os = "linux")'.dependencies]
-vibe_core = { path = "../../core", features = [] }
 openssl = { version = "0.10.68", features = ["vendored"] }
 
 # Windows
 [target.'cfg(windows)'.dependencies]
-vibe_core = { path = "../../core", features = [] }
 winreg = "0.52.0"
-# Used to attach to console
-windows = { version = "0.58.0", features = [
-	"Win32_System_Console",
-	"Win32_Foundation",
-] }
-# Used to check that vulkan works
-ash = { version = "0.38.0", optional = true }
-# Used to redirect stdout/stderr from c++ to the attached console. otherwise whisper errors won't show
-libc = "0.2.158"
-libc-stdhandle = "=0.1.0"
-
 
 # macOS
 [target.'cfg(target_os = "macos")'.dependencies]
-vibe_core = { path = "../../core", features = ["coreml", "metal"] }
 cocoa = "0.26.0"
 screencapturekit = { version = "0.2.8", features = ["ci"] }
 screencapturekit-sys = "0.2.8"
 objc_id = "0.1"
 core-graphics-helmer-fork = "0.24.0"
-
-[features]
-default = []
-cuda = ["vibe_core/cuda"]
-coreml = ["vibe_core/coreml"]
-metal = ["vibe_core/metal"]
-openblas = ["vibe_core/openblas"]
-rocm = ["vibe_core/rocm"]
-vulkan = ["vibe_core/vulkan", "dep:ash"]
-server = ["dep:utoipa", "dep:utoipa-swagger-ui", "dep:axum", "vibe_core/server"]
