+++ desktop/src-tauri/build.rs
@@ -41,29 +41,6 @@ fn copy_locales() {
     copy_folder(src_locales.as_path(), &target_locales);
 }
 
-fn extract_whisper_env() {
-    println!("cargo:rerun-if-env-changed=WHISPER_NO_AVX");
-    println!("cargo:rerun-if-env-changed=WHISPER_NO_AVX2");
-    println!("cargo:rerun-if-env-changed=WHISPER_NO_FMA");
-    println!("cargo:rerun-if-env-changed=WHISPER_NO_F16C");
-    println!(
-        "cargo:rustc-env=WHISPER_NO_AVX={}",
-        std::env::var("WHISPER_NO_AVX").unwrap_or_default().trim()
-    );
-    println!(
-        "cargo:rustc-env=WHISPER_NO_AVX2={}",
-        std::env::var("WHISPER_NO_AVX2").unwrap_or_default().trim()
-    );
-    println!(
-        "cargo:rustc-env=WHISPER_NO_FMA={}",
-        std::env::var("WHISPER_NO_FMA").unwrap_or_default().trim()
-    );
-    println!(
-        "cargo:rustc-env=WHISPER_NO_F16C={}",
-        std::env::var("WHISPER_NO_F16C").unwrap_or_default().trim()
-    );
-}
-
 fn main() {
     let hash = commit_hash();
     println!("cargo:rerun-if-env-changed=COMMIT_HASH");
@@ -91,6 +68,5 @@ fn main() {
     );
 
     copy_locales();
-    extract_whisper_env();
     tauri_build::build();
 }
