+++ desktop/src-tauri/src/cmd/mod.rs
@@ -17,28 +17,9 @@ use tauri::{
 use tauri::{Emitter, Listener, State};
 use tauri_plugin_store::StoreExt;
 use tokio::sync::Mutex;
-use vibe_core::get_vibe_temp_folder;
-use vibe_core::transcript::Segment;
 use vibe_core::transcript::Transcript;
 pub mod audio;
 pub mod ytdlp;
-
-/// Return true if there's internet connection
-/// timeout in ms
-#[tauri::command]
-pub async fn is_online(timeout: Option<u64>) -> Result<bool> {
-    let timeout = std::time::Duration::from_millis(timeout.unwrap_or(2000));
-    let targets = ["1.1.1.1:80", "1.1.1.1:53", "8.8.8.8:53", "8.8.8.8:80"];
-
-    let tasks = targets.iter().map(|addr| async move {
-        tokio::time::timeout(timeout, tokio::net::TcpStream::connect(addr))
-            .await
-            .map(|res| res.is_ok())
-            .unwrap_or(false)
-    });
-
-    Ok(futures::future::join_all(tasks).await.into_iter().any(|res| res))
-}
 
 fn set_progress_bar(app_handle: &tauri::AppHandle, progress: Option<f64>) -> Result<()> {
     let window = app_handle.get_webview_window("main").context("get window")?;
@@ -369,17 +350,6 @@ pub fn get_path_dst(src: String, suffix: String) -> Result<String> {
 }
 
 #[tauri::command]
-#[cfg(windows)]
-pub fn set_high_gpu_preference(mode: bool) -> Result<()> {
-    if mode {
-        crate::gpu_preference::set_gpu_preference_high()?;
-    } else {
-        crate::gpu_preference::remove_gpu_preference()?;
-    }
-    Ok(())
-}
-
-#[tauri::command]
 pub fn get_save_path(src_path: PathBuf, target_ext: &str) -> Result<Value> {
     // Get the file stem (filename without extension)
     let stem = src_path.file_stem().and_then(|s| s.to_str()).unwrap_or_default();
@@ -409,22 +379,6 @@ pub async fn open_path(path: PathBuf) -> Result<()> {
         open::that(path)?;
     }
     Ok(())
-}
-
-#[tauri::command]
-pub fn get_cuda_version() -> String {
-    env!("CUDA_VERSION").to_string()
-}
-
-#[tauri::command]
-pub fn get_rocm_version() -> String {
-    env!("ROCM_VERSION").to_string()
-}
-
-#[tauri::command]
-pub fn is_avx2_enabled() -> bool {
-    #[allow(clippy::comparison_to_empty)]
-    return env!("WHISPER_NO_AVX") != "ON";
 }
 
 #[tauri::command]
@@ -466,43 +420,6 @@ pub async fn load_model(
 #[allow(clippy::comparison_to_empty)]
 pub fn is_portable() -> bool {
     env!("WINDOWS_PORTABLE") == "1"
-}
-
-#[tauri::command]
-pub fn check_vulkan() -> Result<()> {
-    #[cfg(all(feature = "vulkan", windows))]
-    {
-        use ash::vk;
-        unsafe {
-            let entry = match ash::Entry::load() {
-                Ok(e) => e,
-                Err(e) => {
-                    tracing::error!("Failed to load Vulkan entry: {:?}", e);
-                    return Err(e.into());
-                }
-            };
-
-            let app_desc = vk::ApplicationInfo::default().api_version(vk::make_api_version(0, 1, 0, 0));
-            let instance_desc = vk::InstanceCreateInfo::default().application_info(&app_desc);
-
-            let instance = match entry.create_instance(&instance_desc, None) {
-                Ok(inst) => inst,
-                Err(e) => {
-                    tracing::error!("Failed to create Vulkan instance: {:?}", e);
-                    return Err(e.into());
-                }
-            };
-
-            instance.destroy_instance(None);
-            tracing::debug!("Vulkan support is successfully checked and working.");
-        }
-        Ok(())
-    }
-    #[cfg(not(all(feature = "vulkan", windows)))]
-    {
-        tracing::debug!("Vulkan check skipped on this platform");
-        Ok(())
-    }
 }
 
 #[tauri::command]
@@ -554,46 +471,4 @@ pub fn get_logs(app_handle: tauri::AppHandle) -> Result<String> {
     let path = crate::logging::get_log_path(&app_handle)?;
     let content = std::fs::read_to_string(path)?;
     Ok(content)
-}
-
-#[tauri::command]
-pub fn is_crashed_recently() -> bool {
-    tracing::debug!("checking path {}", get_vibe_temp_folder().join("crash.txt").display());
-    get_vibe_temp_folder().join("crash.txt").exists()
-}
-
-#[tauri::command]
-pub fn rename_crash_file() -> Result<()> {
-    std::fs::rename(
-        get_vibe_temp_folder().join("crash.txt"),
-        // TODO: save all crashed?
-        get_vibe_temp_folder().join("crash.1.txt"),
-    )
-    .context("Can't delete file")
-}
-
-#[tauri::command]
-pub fn get_cargo_features() -> Vec<String> {
-    let mut enabled_features = Vec::new();
-
-    if cfg!(feature = "cuda") {
-        enabled_features.push("cuda".to_string());
-    }
-    if cfg!(feature = "coreml") {
-        enabled_features.push("coreml".to_string());
-    }
-    if cfg!(feature = "metal") {
-        enabled_features.push("metal".to_string());
-    }
-    if cfg!(feature = "openblas") {
-        enabled_features.push("openblas".to_string());
-    }
-    if cfg!(feature = "vulkan") {
-        enabled_features.push("vulkan".to_string());
-    }
-    if cfg!(feature = "rocm") {
-        enabled_features.push("rocm".to_string());
-    }
-
-    enabled_features
 }
