+++ desktop/src-tauri/src/screen_capture_kit.rs
@@ -12,8 +12,6 @@ use std::io::Write;
 use std::ops::Deref;
 use std::path::PathBuf;
 use std::process::{Command, Stdio};
-use vibe_core::audio::find_ffmpeg_path;
-use vibe_core::get_vibe_temp_folder;
 
 use crate::utils::LogError;
