+++ server/Cargo.toml
@@ -32,6 +32,18 @@ serde_json = { workspace = true }
 futures-util = "0.3.30"
 pyannote-rs = "0.2.7"
 
+# Used to attach to console
+windows = { version = "0.58.0", features = [
+	"Win32_System_Console",
+	"Win32_Foundation",
+] }
+
+# Used to check that vulkan works
+ash = { version = "0.38.0", optional = true }
+# Used to redirect stdout/stderr from c++ to the attached console. otherwise whisper errors won't show
+libc = "0.2.158"
+libc-stdhandle = "=0.1.0"
+
 # Server
 utoipa = { version = "4.2.3", features = ["axum_extras"], optional = true }
