+++ server/src/gpu_preference.rs
