+++ server/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod audio;
+pub mod config;
+pub mod downloader;
+pub mod transcribe;
+pub mod transcript;
+
+#[cfg(test)]
+mod test;
