+++ server/src/panic_hook.rs
